requirements.txt
# Core data processing
pandas>=2.0.0
numpy>=1.24.0

# API clients and HTTP
requests>=2.31.0
openai>=1.0.0  # OpenAI-compatible client for OpenRouter

# Configuration
python-dotenv>=1.0.0
pyyaml>=6.0

# CLI
click>=8.1.0

# TUI Framework
textual>=0.50.0
rich>=13.0.0

# Web UI
flask>=3.0.0

# Utilities
python-dateutil>=2.8.0
